TASK-29 Buttons extracted to the EditToolbarComponent
